refactor(pr-context): split guard tests
